edit welcome.vue

+++ database/migrations/2025_08_12_103038_drop_date_from_leaves_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('leaves', function (Blueprint $table) {
-            $table->dropColumn('date');
-        });
+        if (Schema::hasColumn('leaves', 'date')) {
+            Schema::table('leaves', function (Blueprint $table) {
+                $table->dropColumn('date');
+            });
+        }
     }
 
     /**
