+++ database/migrations/2025_08_12_104810_add_date_to_leaves_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('leaves', function (Blueprint $table) {
-            $table->dropColumn('date');
-        });
+        if (Schema::hasColumn('leaves', 'date')) {
+            Schema::table('leaves', function (Blueprint $table) {
+                $table->dropColumn('date');
+            });
+        }
     }
 };
